requirements.txt
numpy>=1.20.3
pandas>=1.3.1
matplotlib>=3.4.2
seaborn>=0.11.2
statsmodels>=0.12.0
scikit-learn>=1.0.1
grid_strategy==0.0.1
pyyaml>=6.0
UpSetPlot>=0.5.0
matplotlib-venn==0.11.6
scipy>=1.7.3
hdbscan==0.8.27
scikit-learn-extra>=0.2.0
xlmhg>=2.5.4
numba>=0.54.1
pairwisedist>=1.2.0
requests>=2.24.0
pydot>=1.4.2
joblib>=1.1.0
tqdm>=4.62.3
appdirs>=1.4.0
